Update README.md
# MaaSAlgorithm

**Dataset
The dataset used in this project is located in the code0615/instance folder. 

**Algorithms
*MP_V1.py
The MP_V1 algorithm is the main function for the one-stage algorithm, which includes transport resource allocation and rider routing. 

*Hopcroft_Karp_Algo_dict.py
The Hopcroft_Karp_Algo_dict function is responsible for fleet scheduling. 

*output_result.py
The output_result function is the main function that produces the final output.